requirements.txt
opencv-python
numpy
gTTS
RPi.GPIO
readchar